Add a shared AudioParamDescriptor type under src/types. Replace the duplicate descriptor interface in param-types.ts with an import. Update sample-player descriptors to use the shared type. Preserve literal parameter names with as const satisfies AudioParamDescriptor[].
