+++ packages/audiolib/src/nodes/instruments/Sample/param-defaults.ts
@@ -6,6 +6,7 @@ import {
   SamplePlayerParamKey,
 } from '@/worklets';
 
+
 // Worklet's parameterDescriptors are the source of truth for min/max/default/automationRate.
 const fromWorklet = (name: SamplePlayerParamKey) => {
   const descriptor = SAMPLE_PLAYER_PARAM_DESCRIPTORS.find(
