+++ packages/audiolib/src/types/audio-param-descriptor.ts
@@ -1,0 +1,7 @@
+export interface AudioParamDescriptor {
+  name: string;
+  defaultValue?: number;
+  minValue?: number;
+  maxValue?: number;
+  automationRate: 'a-rate' | 'k-rate';
+}
